move CheckItemRequirementsService computation parts to Post and ItemRequirement models
