+++ src/data/entities/item-requirement/item.requirement.entity.ts
@@ -2,6 +2,11 @@ import { ChildEntity, Column } from "typeorm";
 import { ItemRequirementProps } from "./Item.requirement.props";
 import { Requirement } from "../requirement/requirement.entity";
 
+export interface Item {
+  name: string;
+  quantity: number;
+}
+
 @ChildEntity()
 export class ItemRequirement extends Requirement {
   @Column({ type: "int" })
@@ -12,5 +17,12 @@ export class ItemRequirement extends Requirement {
     if (props) {
       this.quantity = props.quantity;
     }
+  }
+
+  isSatifiedBy(item: Item) {
+    const isRequiredItem = this.name === item.name;
+    const hasEnough = this.quantity <= item.quantity;
+
+    return isRequiredItem && hasEnough;
   }
 }
