--- src/services/requirement/item.ts
@@ -1,4 +1,0 @@
-export interface Item {
-  name: string;
-  quantity: number;
-}
